Done Checkbox

+++ checkbox/function.php
@@ -1,23 +1,33 @@
 <?php
 
- class insertdata
-{
-    public function insert($data,$table2,$table1,$col)
+    class insertdata
     {
-        require('connection.php');
-
-        $data = $_POST['check'];
-       
-
-            foreach($data as $value)
-
+        public function insert($data,$table2,$table1,$col)
         {
-            $query = "INSERT INTO $table2  VALUES ('','$value')";
-            $query_run = mysqli_query($conn,$query);
-            $del_sql = "delete from $table1 where $col = '$value'";
-            //echo $del_sql; exit;
-            mysqli_query($conn,$del_sql);
-        }
+            require('connection.php');
+
+            //$data = $_POST['check'];
+            foreach($data as $value)
+            {
+                $query = "INSERT INTO $table2  VALUES (NULL,'$value')";
+                if(mysqli_query($conn, $query))
+                {
+                    //echo "hello";
+                    $query_run = mysqli_query($conn,$query);
+                    $del_sql = "delete from $table1 where $col = '$value'";
+                    if (mysqli_query($conn, $del)) 
+                    {
+                        header("location:form.php");
+                        // echo "Record deleted successfully";
+                    } 
+                    else 
+                    {
+                        echo "Error deleting record: " . mysqli_error($conn);
+                    }
+                }    
+                    //echo $del_sql; exit;
+                    mysqli_query($conn,$del_sql);
+            }
             if($query_run)
             {
                 echo "Inserted Successfully";
@@ -26,15 +36,11 @@
             else
             {
                 echo "Data Not Inserted";
-                //header("Location: tableA.php");
-            }
-
-            echo "HELLO PHP";
-            exit;
-
+                //header("Location: form.php");
+            }   
         }
 
-        public function getdata($table1,$table2)
+         public function getdata($table1,$table2)
         {
             include('connection.php');
             $query="SELECT * FROM $table1 ORDER BY checkbox_data ASC";
@@ -48,19 +54,5 @@
             $total['data1']=$data1;
             return $total;
         }
-
-  
-
-        
-
-
-
-
-
     }
-
-    
-    
-
-
 ?>
